2025.11.27 c端菜品展示，修改了一些字段类型 ！停售的菜品和套餐不展示

+++ sky-common/src/main/java/com/sky/context/BaseContext.java
@@ -2,13 +2,13 @@ package com.sky.context;
 
 public class BaseContext {
 
-    public static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
+    public static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
 
-    public static void setCurrentId(Long id) {
+    public static void setCurrentId(Integer id) {
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId() {
+    public static Integer getCurrentId() {
         return threadLocal.get();
     }
