+++ sky-pojo/src/main/java/com/sky/entity/AddressBook.java
@@ -18,10 +18,10 @@ public class AddressBook implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private Integer id;
 
     //用户id
-    private Long userId;
+    private Integer userId;
 
     //收货人
     private String consignee;
