+++ sky-server/src/main/java/com/sky/interceptor/JwtTokenAdminInterceptor.java
@@ -46,7 +46,7 @@ public class JwtTokenAdminInterceptor implements HandlerInterceptor {
         try {
             log.info("jwt校验:{}", token);
             Claims claims = JwtUtil.parseJWT(jwtProperties.getAdminSecretKey(), token);
-            Long empId = Long.valueOf(claims.get(JwtClaimsConstant.EMP_ID).toString());
+            Integer empId = Integer.valueOf(claims.get(JwtClaimsConstant.EMP_ID).toString());
             log.info("当前员工id：", empId);
             BaseContext.setCurrentId(empId);  // 将当前员工的id存入当前线程的ThreadLocal中
             //3、通过，放行
